requirements-client.txt
rich
keyboard
pyclip
numpy
sounddevice
websockets
pypinyin
watchdog
typer
srt
pyinstaller
pystray
Pillow